+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Enh #199: `SendMediaGroup` method now supports `InputMediaLivePhoto` type.
 - New #199: Add `BotAccessSettings` type.
 - New #199: Add `GetManagedBotAccessSettings` method.
+- New #199: Add `SetManagedBotAccessSettings` method.
 
 ## 0.20 April 26, 2026